+++ adminforth/spa/src/spa_types/core.ts
@@ -32,6 +32,7 @@ export type CoreConfig = {
         usernameField: string,
         passwordHashField: string,
         loginBackgroundImage: string,
+        loginBackgroundPosition: string,
         userFullnameField: string,
     },
     emptyFieldPlaceholder?: {